pyproject.toml
[tool.poetry]
name = "pyserve"
version = "0.1.0"
description = ""
authors = ["meirdev <maornet@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
requests = "^2.28.1"
mypy = "^0.982"
pytest-asyncio = "^0.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
